Django_Project/requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2025.4
Django==5.2
django-crispy-forms==2.4
django-debug-toolbar==5.2.0
djangorestframework==3.16.0
pillow==11.2.1
psycopg2-binary==2.9.10
sqlparse==0.5.3
